% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nearest_pos.R
\name{find_nearest_pos}
\alias{find_nearest_pos}
\title{Title}
\usage{
find_nearest_pos(
  lat_arr,
  lon_arr,
  lat,
  lon,
  prec_lat = 0.005,
  prec_lon = 0.005
)
}
\arguments{
\item{lat_arr}{2D Array of latitute in WRF}

\item{lon_arr}{2D array of longitude in WRF}

\item{lat}{Latitude of position in question}

\item{lon}{Longitute of position in qestion}

\item{prec_lat}{Precision for Latitute}

\item{prec_lon}{}
}
\value{

}
\description{
Title
}
