% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pot_temp.R
\name{get_pot_temp_2D}
\alias{get_pot_temp_2D}
\title{Title}
\usage{
get_pot_temp_2D(
  ncfile,
  x_start = 1,
  y_start = 1,
  z_start = 1,
  t_start = 1,
  x_len,
  y_len,
  z_len = 1,
  t_len = 1
)
}
\arguments{
\item{zlayer}{}
}
\value{

}
\description{
Title
}
