% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temp_and_q.R
\name{get_temp_and_q}
\alias{get_temp_and_q}
\title{Title}
\usage{
get_temp_and_q(
  ncfile,
  x_start = 1,
  y_start = 1,
  z_start = 1,
  t_start = 1,
  x_len,
  y_len,
  z_len = 1,
  t_len = 1,
  moisture = TRUE
)
}
\arguments{
\item{moisture}{}
}
\value{

}
\description{
Title
}
