% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wind.R
\name{get_wind}
\alias{get_wind}
\title{Title}
\usage{
get_wind(
  ncfile,
  x_start = 1,
  y_start = 1,
  z_start = 1,
  t_start = 1,
  x_len,
  y_len,
  z_len = 1,
  t_len = 1
)
}
\arguments{
\item{tstep}{}
}
\value{

}
\description{
Title
}
