% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_operators.R
\name{\%and\%}
\alias{\%and\%}
\title{%and% operator similar to python}
\usage{
lhs \%and\% rhs
}
\arguments{
\item{lhs}{Left hand side}

\item{rhs}{Right hand side}
}
\value{
Bool.
}
\description{
%and% operator similar to python
}
\examples{
a <- 2 > 0
b <- TRUE
a \%and\% b
}
