% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_operators.R
\name{\%nor\%}
\alias{\%nor\%}
\title{%nor% operator, just to have it}
\usage{
lhs \%nor\% rhs
}
\arguments{
\item{lhs}{Left hand side}

\item{rhs}{Right hand side}
}
\value{
Bool.
}
\description{
%nor% operator, just to have it
}
\examples{
a <- 2 < 0
b <- FALSE
a \%nor\% b
}
