% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_operators.R
\name{\%or\%}
\alias{\%or\%}
\title{%or% operator similar to python}
\usage{
lhs \%or\% rhs
}
\arguments{
\item{lhs}{Left hand side}

\item{rhs}{Right hand side}
}
\value{
Bool.
}
\description{
%or% operator similar to python
}
\examples{
a <- 2 < 0
b <- TRUE
a \%or\% b
}
