% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_python.R
\name{theme_python}
\alias{theme_python}
\title{Title}
\usage{
theme_python(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/170,
  base_rect_size = base_size/170,
  x_pos_leg = 0.1,
  y_pos_leg = 0.85,
  aspect = 5.5/8
)
}
\arguments{
\item{aspect}{}
}
\value{

}
\description{
Title
}
