% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_python2.R
\name{theme_python2}
\alias{theme_python2}
\title{Title}
\usage{
theme_python2(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/170,
  base_rect_size = base_size/170,
  x_pos_leg = 0.1,
  y_pos_leg = 0.85,
  aspect = 5.5/8
)
}
\arguments{
\item{base_size}{}

\item{base_family}{}

\item{base_line_size}{}

\item{base_rect_size}{}

\item{x_pos_leg}{}

\item{y_pos_leg}{}

\item{aspect}{}
}
\value{

}
\description{
Title
}
